requirements/base.txt
Django==1.5.1
Markdown==2.3.1
Moka==0.0.5
Pillow==2.1.0
Pygments==1.6
South==0.8.2
distribute==0.6.34
dj-database-url==0.2.1
dj-static==0.0.5
django-classy-tags==0.4
django-extensions==1.1.1
django-filter==0.7
django-sekizai==0.7
django-toolbelt==0.0.1
djangorestframework==2.3.9
gunicorn==0.17.4
hamlpy==0.82.2
psycopg2==2.5.1
pygraphviz==1.2
requests==1.2.3
six==1.3.0
static==0.4
wsgiref==0.1.2
